extract simplify into its own module
